requirements.txt
Django==2.2.11
pytz==2019.3
sqlparse==0.3.1
psycopg2-binary==2.8.4
gunicorn==20.0.4
sentry-sdk==0.14.3
djangorestframework==3.11.0
